api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
drf-camel-case==1.0.2
drf-spectacular==0.29.0
et_xmlfile==2.0.0
frozenlist==1.8.0
http_ece==1.2.1
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.7.0
openpyxl==3.1.5
orjson==3.11.5
pillow==12.0.0
propcache==0.4.1
py-vapid==1.9.2
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-decouple==3.8
pywebpush==2.1.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.1
yarl==1.22.0
